requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
celery==5.3.1
redis==4.5.5
sqlalchemy==2.0.15
pillow==9.5.0
requests==2.28.2
psycopg2-binary==2.9.6
python-multipart==0.0.6
cloudinary==1.29.0
gunicorn==21.2.0